order-read-model-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
kafka-python
requests
